refactor: remove unused function
