+++ pkg/versiongetter/github_release_internal_test.go
@@ -1,4 +1,4 @@
-package generate
+package versiongetter
 
 import (
 	"testing"
